requirements.txt
networkx>=2.5
matplotlib>=3.3.2
torch==1.8.1
torchvision==0.9.1
torchaudio===0.8.1
PyYAML==5.4.1
graphviz==0.16